Move actor-group and actor related principals to own resource
