+++ libs/@local/graph/api/src/rest/actor_group.rs
@@ -9,7 +9,7 @@ use axum::{
     response::Response,
     routing::{get, post},
 };
-use error_stack::Report;
+use error_stack::{Report, ResultExt as _};
 use hash_graph_authorization::{
     AuthorizationApi as _, AuthorizationApiPool,
     backend::{ModifyRelationshipOperation, PermissionAssertion},
@@ -37,42 +37,25 @@ use type_system::principal::{
 };
 use utoipa::{OpenApi, ToSchema};
 
+use super::{OpenApiQuery, QueryLogger};
 use crate::rest::{AuthenticatedUserHeader, PermissionResponse, status::report_to_response};
 
 #[derive(OpenApi)]
 #[openapi(
     paths(
-        get_actor_group_role_assignments,
-        has_actor_group_role,
-        assign_actor_group_role,
-        unassign_actor_group_role,
-
-        get_web_by_id,
-        get_web_by_shortname,
-        create_org_web,
         check_web_permission,
         modify_web_authorization_relationships,
         get_web_authorization_relationships,
-
-        get_team_by_name,
+        check_account_group_permission,
+        get_actor_group_relations,
     ),
     components(
         schemas(
-            GetWebResponse,
-            CreateOrgWebParams,
-            CreateWebResponse,
-
-            ActorGroupEntityUuid,
-            ActorGroupId,
-            RoleName,
-            RoleAssignmentStatus,
-            RoleUnassignmentStatus,
             AccountGroupPermission,
             AccountGroupRelationAndSubject,
             AccountGroupMemberSubject,
             AccountGroupAdministratorSubject,
 
-            WebId,
             WebRelationAndSubject,
             WebPermission,
             WebOwnerSubject,
@@ -83,9 +66,6 @@ use crate::rest::{AuthenticatedUserHeader, PermissionResponse, status::report_to
             WebPropertyTypeViewerSubject,
             WebDataTypeViewerSubject,
             ModifyWebAuthorizationRelationship,
-
-            GetTeamResponse,
-            TeamId,
         ),
     ),
     tags(
@@ -106,184 +86,33 @@ impl ActorGroupResource {
             "/actor-groups",
             Router::new()
                 .nest(
-                    "/:actor_group_id",
+                    "/:actor_group_id/",
                     Router::new()
                         .route(
-                            "/roles/:role/actors",
-                            get(get_actor_group_role_assignments::<S, A>),
+                            "/permissions/:permission",
+                            get(check_account_group_permission::<A>),
                         )
-                        .route(
-                            "/roles/:role/actors/:actor_id",
-                            get(has_actor_group_role::<S, A>)
-                                .post(assign_actor_group_role::<S, A>)
-                                .delete(unassign_actor_group_role::<S, A>),
-                        ),
+                        .route("/relations", get(get_actor_group_relations::<A>)),
                 )
                 .nest(
                     "/webs",
                     Router::new()
-                        .route("/", post(create_org_web::<S, A>))
                         .nest(
                             "/:web_id",
                             Router::new()
-                                .route("/", get(get_web_by_id::<S, A>))
                                 .route("/permissions/:permission", get(check_web_permission::<A>))
                                 .route(
                                     "/relationships",
                                     get(get_web_authorization_relationships::<A>),
                                 ),
                         )
-                        .route("/shortname/:shortname", get(get_web_by_shortname::<S, A>))
                         .route(
                             "/relationships",
                             post(modify_web_authorization_relationships::<A>),
                         ),
-                )
-                .nest(
-                    "/teams",
-                    Router::new().route("/name/:name", get(get_team_by_name::<S, A>)),
                 ),
         )
     }
-}
-
-#[utoipa::path(
-    post,
-    path = "/actor-groups/webs",
-    request_body = CreateOrgWebParams,
-    tag = "Web",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-    ),
-    responses(
-        (status = 200, content_type = "application/json", description = "The web was created successfully", body = CreateWebResponse),
-
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn create_org_web<S, A>(
-    AuthenticatedUserHeader(actor_id): AuthenticatedUserHeader,
-    authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-    Json(params): Json<CreateOrgWebParams>,
-) -> Result<Json<CreateWebResponse>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    let authorization_api = authorization_api_pool
-        .acquire()
-        .await
-        .map_err(report_to_response)?;
-
-    let mut store = store_pool
-        .acquire(authorization_api, temporal_client.0)
-        .await
-        .map_err(report_to_response)?;
-
-    store
-        .create_org_web(actor_id, params)
-        .await
-        .map(Json)
-        .map_err(report_to_response)
-}
-
-#[utoipa::path(
-    get,
-    path = "/actor-groups/webs/{web_id}",
-    tag = "Web",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("web_id" = WebId, Path, description = "The ID of the web to retrieve"),
-    ),
-    responses(
-        (status = 200, content_type = "application/json", description = "The web was retrieved successfully", body = GetWebResponse),
-
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn get_web_by_id<S, A>(
-    AuthenticatedUserHeader(actor_id): AuthenticatedUserHeader,
-    Path(web_id): Path<WebId>,
-    authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-) -> Result<Json<Option<GetWebResponse>>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    store_pool
-        .acquire(
-            authorization_api_pool
-                .acquire()
-                .await
-                .map_err(report_to_response)?,
-            temporal_client.0,
-        )
-        .await
-        .map_err(report_to_response)?
-        .get_web_by_id(actor_id, web_id)
-        .await
-        .map_err(report_to_response)
-        .map(Json)
-}
-
-#[utoipa::path(
-    get,
-    path = "/actor-groups/webs/shortname/{shortname}",
-    tag = "Web",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("shortname" = String, Path, description = "The shortname of the web to retrieve"),
-    ),
-    responses(
-        (status = 200, content_type = "application/json", description = "The web was retrieved successfully", body = GetWebResponse),
-
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn get_web_by_shortname<S, A>(
-    AuthenticatedUserHeader(actor_id): AuthenticatedUserHeader,
-    Path(shortname): Path<String>,
-    authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-) -> Result<Json<Option<GetWebResponse>>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    store_pool
-        .acquire(
-            authorization_api_pool
-                .acquire()
-                .await
-                .map_err(report_to_response)?,
-            temporal_client.0,
-        )
-        .await
-        .map_err(report_to_response)?
-        .get_web_by_shortname(actor_id, &shortname)
-        .await
-        .map_err(report_to_response)
-        .map(Json)
 }
 
 #[utoipa::path(
@@ -454,257 +283,111 @@ where
 
 #[utoipa::path(
     get,
-    path = "/actor-groups/{actor_group_id}/roles/{role_name}/actors",
-    tag = "ActorGroup",
+    path = "/actor_groups/{actor_group_id}/permissions/{permission}",
+    tag = "Actor Group",
     params(
         ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to get the assignments for"),
-        ("role_name" = RoleName, Path, description = "The role to be checked"),
+        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to check if the actor has the permission"),
+        ("permission" = AccountGroupPermission, Path, description = "The permission to check for"),
     ),
     responses(
-        (status = 200, content_type = "application/json", description = "Whether the actor is assigned the role", body = Vec<ActorEntityUuid>),
-        (status = 404, content_type = "application/json", description = "The team was not found"),
+        (status = 200, body = PermissionResponse, description = "Information if the actor can add an owner"),
 
-        (status = 500, description = "Store error occurred"),
+        (status = 500, description = "Internal error occurred"),
     )
 )]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn get_actor_group_role_assignments<S, A>(
+#[tracing::instrument(level = "info", skip(authorization_api_pool))]
+async fn check_account_group_permission<A>(
     AuthenticatedUserHeader(actor): AuthenticatedUserHeader,
-    Path((actor_group_id, role_name)): Path<(ActorGroupEntityUuid, RoleName)>,
+    Path((actor_group_id, permission)): Path<(ActorGroupEntityUuid, AccountGroupPermission)>,
     authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-) -> Result<Json<Vec<ActorEntityUuid>>, Response>
+    mut query_logger: Option<Extension<QueryLogger>>,
+) -> Result<Json<PermissionResponse>, Response>
 where
-    S: StorePool + Send + Sync,
     A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
 {
-    let authorization_api = authorization_api_pool
-        .acquire()
-        .await
-        .map_err(report_to_response)?;
+    if let Some(query_logger) = &mut query_logger {
+        query_logger.capture(
+            actor,
+            OpenApiQuery::CheckAccountGroupPermission {
+                actor_group_id,
+                permission,
+            },
+        );
+    }
 
-    let mut store = store_pool
-        .acquire(authorization_api, temporal_client.0)
-        .await
-        .map_err(report_to_response)?;
-
-    store
-        .get_role_assignments(actor_group_id, role_name)
-        .await
-        .map_err(report_to_response)
-        .map(Json)
-}
-
-#[utoipa::path(
-    get,
-    path = "/actor-groups/{actor_group_id}/roles/{role_name}/actors/{actor_id}",
-    tag = "ActorGroup",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to check the role against"),
-        ("role_name" = RoleName, Path, description = "The role to be checked"),
-        ("actor_id" = ActorEntityUuid, Path, description = "The ID of the actor to be checked"),
-    ),
-    responses(
-        (status = 200, content_type = "application/json", description = "Whether the actor is assigned the role", body = bool),
-        (status = 404, content_type = "application/json", description = "The team was not found"),
-
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn has_actor_group_role<S, A>(
-    AuthenticatedUserHeader(actor): AuthenticatedUserHeader,
-    Path((actor_group_id, role_name, actor_id)): Path<(
-        ActorGroupEntityUuid,
-        RoleName,
-        ActorEntityUuid,
-    )>,
-    authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-) -> Result<Json<bool>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    let authorization_api = authorization_api_pool
-        .acquire()
-        .await
-        .map_err(report_to_response)?;
-
-    let mut store = store_pool
-        .acquire(authorization_api, temporal_client.0)
-        .await
-        .map_err(report_to_response)?;
-
-    Ok(Json(
-        store
-            .is_assigned(actor_id, actor_group_id)
+    let response = Ok(Json(PermissionResponse {
+        has_permission: authorization_api_pool
+            .acquire()
             .await
             .map_err(report_to_response)?
-            .is_some_and(|role| role == role_name),
-    ))
-}
-
-#[utoipa::path(
-    post,
-    path = "/actor-groups/{actor_group_id}/roles/{role_name}/actors/{actor_id}",
-    tag = "ActorGroup",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to add the member to"),
-        ("role_name" = RoleName, Path, description = "The role to assign to the actor"),
-        ("actor_id" = ActorEntityUuid, Path, description = "The ID of the actor to add to the group"),
-    ),
-    responses(
-        (status = 200, body = RoleAssignmentStatus, description = "The actor group member was added"),
-
-        (status = 403, description = "Permission denied"),
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn assign_actor_group_role<S, A>(
-    AuthenticatedUserHeader(actor): AuthenticatedUserHeader,
-    Path((actor_group_id, role_name, actor_id)): Path<(
-        ActorGroupEntityUuid,
-        RoleName,
-        ActorEntityUuid,
-    )>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-    authorization_api_pool: Extension<Arc<A>>,
-) -> Result<Json<RoleAssignmentStatus>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    store_pool
-        .acquire(
-            authorization_api_pool
-                .acquire()
-                .await
-                .map_err(report_to_response)?,
-            temporal_client.0,
-        )
-        .await
-        .map_err(report_to_response)?
-        .assign_role(actor, actor_id, actor_group_id, role_name)
-        .await
-        .map(Json)
-        .map_err(report_to_response)
-}
-
-#[utoipa::path(
-    delete,
-    path = "/actor-groups/{actor_group_id}/roles/{role_name}/actors/{actor_id}",
-    tag = "ActorGroup",
-    params(
-        ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to remove the member from"),
-        ("role_name" = RoleName, Path, description = "The role to remove from the actor"),
-        ("actor_id" = ActorEntityUuid, Path, description = "The ID of the actor to remove from the group")
-    ),
-    responses(
-        (status = 200, body = RoleUnassignmentStatus, description = "The actor group member was removed"),
-
-        (status = 403, description = "Permission denied"),
-        (status = 500, description = "Store error occurred"),
-    )
-)]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn unassign_actor_group_role<S, A>(
-    AuthenticatedUserHeader(actor): AuthenticatedUserHeader,
-    Path((actor_group_id, role_name, actor_id)): Path<(
-        ActorGroupEntityUuid,
-        RoleName,
-        ActorEntityUuid,
-    )>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-    authorization_api_pool: Extension<Arc<A>>,
-) -> Result<Json<RoleUnassignmentStatus>, Response>
-where
-    S: StorePool + Send + Sync,
-    A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
-{
-    store_pool
-        .acquire(
-            authorization_api_pool
-                .acquire()
-                .await
-                .map_err(report_to_response)?,
-            temporal_client.0,
-        )
-        .await
-        .map_err(report_to_response)?
-        .unassign_role(actor, actor_id, actor_group_id, role_name)
-        .await
-        .map(Json)
-        .map_err(report_to_response)
+            .check_account_group_permission(
+                actor,
+                permission,
+                actor_group_id,
+                Consistency::FullyConsistent,
+            )
+            .await
+            .attach_printable(
+                "Could not check if permission on the actor group is granted to the specified \
+                 actor",
+            )
+            .map_err(report_to_response)?
+            .has_permission,
+    }));
+    if let Some(query_logger) = &mut query_logger {
+        query_logger.send().await.map_err(report_to_response)?;
+    }
+    response
 }
 
 #[utoipa::path(
     get,
-    path = "/actor-groups/teams/name/{name}",
-    tag = "Team",
+    path = "/actor_groups/{actor_group_id}/relations",
+    tag = "Actor Group",
     params(
         ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
-        ("name" = String, Path, description = "The ID of the team to retrieve"),
+        ("actor_group_id" = ActorGroupEntityUuid, Path, description = "The ID of the actor group to get relations from"),
     ),
     responses(
-        (status = 200, content_type = "application/json", description = "The team was retrieved successfully", body = Option<GetTeamResponse>),
-
-        (status = 500, description = "Store error occurred"),
+        (status = 200, body = Vec<AccountGroupRelationAndSubject>, description = "List of members and administrators of the actor group"),
+        (status = 403, description = "Permission denied"),
+        (status = 500, description = "Internal error occurred"),
     )
 )]
-#[tracing::instrument(
-    level = "info",
-    skip(store_pool, authorization_api_pool, temporal_client)
-)]
-async fn get_team_by_name<S, A>(
-    AuthenticatedUserHeader(actor): AuthenticatedUserHeader,
-    Path(name): Path<String>,
+#[tracing::instrument(level = "info", skip(authorization_api_pool))]
+async fn get_actor_group_relations<A>(
+    AuthenticatedUserHeader(actor_id): AuthenticatedUserHeader,
+    Path(actor_group_id): Path<ActorGroupEntityUuid>,
     authorization_api_pool: Extension<Arc<A>>,
-    temporal_client: Extension<Option<Arc<TemporalClient>>>,
-    store_pool: Extension<Arc<S>>,
-) -> Result<Json<Option<GetTeamResponse>>, Response>
+    mut query_logger: Option<Extension<QueryLogger>>,
+) -> Result<Json<Vec<AccountGroupRelationAndSubject>>, Response>
 where
-    S: StorePool + Send + Sync,
     A: AuthorizationApiPool + Send + Sync,
-    for<'p, 'a> S::Store<'p, A::Api<'a>>: PrincipalStore,
 {
-    store_pool
-        .acquire(
-            authorization_api_pool
-                .acquire()
-                .await
-                .map_err(report_to_response)?,
-            temporal_client.0,
-        )
+    if let Some(query_logger) = &mut query_logger {
+        query_logger.capture(
+            actor_id,
+            OpenApiQuery::GetActorGroupRelations { actor_group_id },
+        );
+    }
+
+    let authorization_api = authorization_api_pool
+        .acquire()
         .await
-        .map_err(report_to_response)?
-        .get_team_by_name(actor, &name)
+        .map_err(report_to_response)?;
+
+    // Get relations of the actor group
+    let result = authorization_api
+        .get_account_group_relations(actor_group_id, Consistency::FullyConsistent)
         .await
+        .attach_printable("Could not get actor group relations")
         .map_err(report_to_response)
-        .map(Json)
+        .map(Json);
+
+    if let Some(query_logger) = &mut query_logger {
+        query_logger.send().await.map_err(report_to_response)?;
+    }
+
+    result
 }
