+++ libs/@local/graph/api/src/rest/permissions.rs
@@ -17,6 +17,7 @@ use hash_graph_authorization::{
 };
 use hash_graph_store::pool::StorePool;
 use hash_temporal_client::TemporalClient;
+use http::StatusCode;
 use type_system::principal::actor::ActorId;
 use utoipa::OpenApi;
 
@@ -28,9 +29,11 @@ use crate::rest::{AuthenticatedUserHeader, json::Json, status::report_to_respons
         get_policy_by_id,
         query_policies,
         resolve_policies_for_actor,
+
+        seed_system_policies,
     ),
     tags(
-        (name = "Permissions", description = "Permission management API")
+        (name = "Permission", description = "Permission management API")
     )
 )]
 pub(crate) struct PermissionResource;
@@ -49,7 +52,8 @@ impl PermissionResource {
             Router::new()
                 .route("/:policy_id", get(get_policy_by_id::<S, A>))
                 .route("/query", post(query_policies::<S, A>))
-                .route("/resolve/actor", post(resolve_policies_for_actor::<S, A>)),
+                .route("/resolve/actor", post(resolve_policies_for_actor::<S, A>))
+                .route("/seed", get(seed_system_policies::<S, A>)),
         )
     }
 }
@@ -57,7 +61,7 @@ impl PermissionResource {
 #[utoipa::path(
     get,
     path = "/policies/{policy_id}",
-    tag = "Policy",
+    tag = "Permission",
     params(
         ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
         ("policy_id" = Uuid, Path, description = "The ID of the policy to find"),
@@ -104,7 +108,7 @@ where
     post,
     path = "/policies/query",
     request_body = Value,
-    tag = "Policy",
+    tag = "Permission",
     params(
         ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
     ),
@@ -150,7 +154,7 @@ where
     post,
     path = "/policies/resolve/actor",
     request_body = Value,
-    tag = "Policy",
+    tag = "Permission",
     params(
         ("X-Authenticated-User-Actor-Id" = ActorEntityUuid, Header, description = "The ID of the actor which is used to authorize the request"),
     ),
@@ -190,4 +194,44 @@ where
         .await
         .map_err(report_to_response)
         .map(Json)
+}
+
+#[utoipa::path(
+    get,
+    path = "/policies/seed",
+    tag = "Permission",
+    responses(
+        (status = 204, content_type = "application/json", description = "The system policies were created successfully"),
+
+        (status = 500, description = "Store error occurred"),
+    )
+)]
+#[tracing::instrument(
+    level = "info",
+    skip(store_pool, authorization_api_pool, temporal_client)
+)]
+async fn seed_system_policies<S, A>(
+    authorization_api_pool: Extension<Arc<A>>,
+    temporal_client: Extension<Option<Arc<TemporalClient>>>,
+    store_pool: Extension<Arc<S>>,
+) -> Result<StatusCode, Response>
+where
+    S: StorePool + Send + Sync,
+    A: AuthorizationApiPool + Send + Sync,
+    for<'p, 'a> S::Store<'p, A::Api<'a>>: PolicyStore,
+{
+    store_pool
+        .acquire(
+            authorization_api_pool
+                .acquire()
+                .await
+                .map_err(report_to_response)?,
+            temporal_client.0,
+        )
+        .await
+        .map_err(report_to_response)?
+        .seed_system_policies()
+        .await
+        .map_err(report_to_response)?;
+    Ok(StatusCode::NO_CONTENT)
 }
